fix: checkin_out BIGINT (Telegram user_id INT4 overflow)
Migration 032 jadvalni INTEGER (INT4) bilan yaratdi.
Mashrab uid=7888864785 — bu int32 max (2147483647)dan katta!
Natijada /tashrif/tarix → asyncpg.DataError: value out of int32 range.

Tuzatish:
- Migration 033: ALTER user_id/klient_id/id → BIGINT
- guards_v2.py CHECK_IN_OUT_MIGRATION ham yangilandi
  (yangi installlarda to'g'ri yaratilsin)

+++ shared/services/guards_v2.py
@@ -85,8 +85,8 @@ async def qoldiq_qaytarish(conn, uid: int, sotuv_id: int) -> dict:
 CHECK_IN_OUT_MIGRATION = """
 CREATE TABLE IF NOT EXISTS checkin_out (
     id SERIAL PRIMARY KEY,
-    user_id INTEGER NOT NULL,
-    klient_id INTEGER NOT NULL,
+    user_id BIGINT NOT NULL,
+    klient_id BIGINT NOT NULL,
     turi VARCHAR(10) NOT NULL,  -- checkin / checkout
     latitude DOUBLE PRECISION,
     longitude DOUBLE PRECISION,
